move PhotoNuclearModel base class into PhotoNuclearModels library
