+++ SimCore/include/SimCore/PhotoNuclearModels/BertiniNothingHardModel.h
@@ -9,7 +9,7 @@
 #include <G4ProcessManager.hh>
 
 #include "Framework/Configure/Parameters.h"
-#include "SimCore/PhotoNuclearModel.h"
+#include "SimCore/PhotoNuclearModels/PhotoNuclearModel.h"
 #include "SimCore/PhotoNuclearModels/BertiniEventTopologyProcess.h"
 namespace simcore {
