+++ SimCore/include/SimCore/PhotoNuclearModels/NoPhotoNuclearModel.h
@@ -4,7 +4,7 @@
 #include <G4ProcessManager.hh>
 
 #include "Framework/Configure/Parameters.h"
-#include "SimCore/PhotoNuclearModel.h"
+#include "SimCore/PhotoNuclearModels/PhotoNuclearModel.h"
 
 namespace simcore {
