+++ SimCore/src/SimCore/PhotoNuclearModels/PhotoNuclearModel.cxx
@@ -1,4 +1,4 @@
-#include "SimCore/PhotoNuclearModel.h"
+#include "SimCore/PhotoNuclearModels/PhotoNuclearModel.h"
 
 namespace simcore {
 void PhotoNuclearModel::removeExistingModel(G4ProcessManager* processManager) {
